test(Go SDK): SC-692 Setup Go SDK benchmarks
