+++ go/packages/slslambda/fixtures/bare/go.mod
@@ -1,0 +1,7 @@
+module bare
+
+go 1.20
+
+require github.com/aws/aws-lambda-go v1.38.0
+
+require github.com/stretchr/testify v1.8.0 // indirect
