+++ go/packages/slslambda/fixtures/bare/main.go
@@ -1,0 +1,11 @@
+package main
+
+import "github.com/aws/aws-lambda-go/lambda"
+
+func main() {
+	lambda.Start(handle)
+}
+
+func handle() (string, error) {
+	return "ok", nil
+}
