+++ node/test/go/aws-lambda-sdk/integration.test.js
@@ -12,9 +12,8 @@ const getProcessFunction = require('../../lib/get-process-function');
 const resolveOutcomeEnumValue = require('../../utils/resolve-outcome-enum-value');
 const resolveNanosecondsTimestamp = require('../../utils/resolve-nanoseconds-timestamp');
 const normalizeEvents = require('../../utils/normalize-events');
-const spawn = require('child-process-ext/spawn');
 const resolveTestVariantsConfig = require('../../lib/resolve-test-variants-config');
-const resolveDirZipBuffer = require('../../utils/resolve-dir-zip-buffer');
+const buildZipFile = require('./lib/build-zip-file');
 
 const slslambdaDirname = path.resolve(__dirname, '../../../../go/packages/slslambda');
 const fixturesDirname = path.resolve(slslambdaDirname, 'fixtures');
@@ -29,22 +28,6 @@ for (const name of ['TEST_EXTERNAL_LAYER_FILENAME']) {
 describe('Go: integration', function () {
   this.timeout(120000);
   const coreConfig = {};
-
-  function buildZipFile(functionName, architecture) {
-    return async () => {
-      const script = path.resolve(slslambdaDirname, 'scripts', 'build-go-binary.sh');
-      const moduleDir = path.resolve(fixturesDirname, functionName);
-      await spawn(script, [moduleDir, architecture]);
-      const outputDir = path.resolve(moduleDir, 'build', architecture);
-      const zipFile = await resolveDirZipBuffer(outputDir);
-      const isArm64 = architecture === 'arm64';
-      return {
-        Runtime: isArm64 ? 'provided.al2' : 'go1.x',
-        Architectures: isArm64 ? ['arm64'] : ['x86_64'],
-        Code: { ZipFile: zipFile },
-      };
-    };
-  }
 
   const useCasesConfig = new Map([
     [
